requirements.txt
Django==4.2.7
Pillow==10.1.0
django-crispy-forms==2.1
crispy-bootstrap4==2024.1
django-allauth==0.57.0
django-markdownx==4.0.2
django-taggit==5.0.1
django-widget-tweaks==1.5.0
python-decouple==3.8
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0
django-cors-headers==4.3.0
djangorestframework==3.14.0
redis==5.0.1
celery==5.3.4
django-redis==5.4.0
Faker==20.1.0
